requirements.txt
generative-ai-hub-sdk
fastapi
hdbcli
python-dotenv
uvicorn
python-multipart
pandas
Flask